Bootstrap terrain chunks on play

+++ UnityProject/Fantasy_World_GIS/Assets/Scripts/Terrain/TerrainBounds.cs
@@ -5,14 +5,18 @@ namespace Fantasy_World_GIS.Terrain
     /// <summary>
     /// Axis-aligned world-space bounds for a terrain dataset or tile.
     /// X/Y are horizontal GIS coordinates in meters.
+    ///
+    /// Unity JsonUtility reliably serializes floats, so the Unity runtime
+    /// representation stores bounds as float values while the platform
+    /// documentation can still describe exported GIS data at double precision.
     /// </summary>
     [Serializable]
     public class TerrainBounds
     {
-        public double MinX;
-        public double MinY;
-        public double MaxX;
-        public double MaxY;
+        public float MinX;
+        public float MinY;
+        public float MaxX;
+        public float MaxY;
 
         public TerrainBounds()
         {
@@ -20,10 +24,10 @@ namespace Fantasy_World_GIS.Terrain
 
         public TerrainBounds(double minX, double minY, double maxX, double maxY)
         {
-            MinX = minX;
-            MinY = minY;
-            MaxX = maxX;
-            MaxY = maxY;
+            MinX = (float)minX;
+            MinY = (float)minY;
+            MaxX = (float)maxX;
+            MaxY = (float)maxY;
         }
     }
 }
